Add put request Add delete request

+++ controller/user.controller.js
@@ -19,9 +19,22 @@ exports.get=(done)=>{
     })
 };
 
-exports.delete=(id)=>{
-   user.destroy({where:{
-       uid:id
-       }
-       })
+exports.deleteById=(id,done)=>{
+    user.findOne({where:{uid:id}}).then((getdata)=>{
+        if(getdata){
+            user.destroy({where:{uid:id}}).then((data)=>{
+                done(null,data)
+            }).catch((err)=>{
+                console.log(err)
+            })
+        }
+    })
+};
+
+exports.updateById=(id,body,done)=>{
+    user.update(body,{where:{uid:id}}).then((updatedata)=>{
+        done(null,updatedata)
+    }).catch((err)=>{
+        console.log(err)
+    })
 };
